+++ parse.cpp
@@ -105,7 +105,7 @@
          cntSpaces++;
       }
 
-      long n = (char *)(*ppEnd + 1) - pSource;
+      long n = (long)((char *)(*ppEnd + 1) - pSource);
       char *pNew = new char[n];
 
       memcpy(pNew,pSource,n);
